feat: divide photos in details and photopicker sub modules
