+++ .idea/gradle.xml
@@ -20,6 +20,9 @@
             <option value="$PROJECT_DIR$/features/albums/home" />
             <option value="$PROJECT_DIR$/features/albums/shared" />
             <option value="$PROJECT_DIR$/features/photos" />
+            <option value="$PROJECT_DIR$/features/photos/detail" />
+            <option value="$PROJECT_DIR$/features/photos/photopicker" />
+            <option value="$PROJECT_DIR$/features/photos/shared" />
             <option value="$PROJECT_DIR$/platform" />
             <option value="$PROJECT_DIR$/platform/database" />
             <option value="$PROJECT_DIR$/platform/database/models" />
