+++ app/build.gradle.kts
@@ -55,7 +55,8 @@ dependencies {
     implementation(projects.features.albums.home)
     implementation(projects.features.albums.create)
     implementation(projects.features.albums.details)
-    implementation(projects.features.photos)
+    implementation(projects.features.photos.detail)
+    implementation(projects.features.photos.photopicker)
 
     implementation(libs.androidX.core.ktx)
     implementation(libs.androidX.lifecycle.runtime)
