+++ app/src/main/java/com/example/mvisample/MviSampleApplication.kt
@@ -6,7 +6,8 @@ import dagger.hilt.android.HiltAndroidApp
 import features.albums.create.di.createAlbumStepDi
 import features.albums.details.di.albumDetailsScreenDi
 import features.albums.home.di.albumHomeStepDi
-import features.photos.di.photoScreenModule
+import features.photos.detail.di.photoDetailsScreenModule
+import features.photos.photopicker.di.photoPickerScreenModule
 
 @HiltAndroidApp
 class MviSampleApplication : Application() {
@@ -17,7 +18,8 @@ class MviSampleApplication : Application() {
             createAlbumStepDi()
             albumHomeStepDi()
             albumDetailsScreenDi()
-            photoScreenModule()
+            photoDetailsScreenModule()
+            photoPickerScreenModule()
         }
     }
