+++ features/photos/detail/src/main/kotlin/features/photos/detail/presentation/PhotoDetailsState.kt
@@ -1,0 +1,8 @@
+package features.photos.detail.presentation
+
+import platform.database.models.models.photo.PhotoModel
+
+internal data class PhotoDetailsState(
+    val photo: PhotoModel = PhotoModel.fixture,
+    val isInitialized: Boolean = false
+)
