+++ features/photos/photopicker/src/main/kotlin/features/photos/photopicker/di/PhotoPickerScreenModule.kt
@@ -1,0 +1,15 @@
+package features.photos.photopicker.di
+
+import cafe.adriel.voyager.core.registry.screenModule
+import features.photos.photopicker.presentation.pickertype.PhotoAddTypeScreen
+import platform.navigation.NavigationProvider
+
+val photoPickerScreenModule = screenModule {
+
+    register<NavigationProvider.Photos.PickPhotoSource> {
+        PhotoAddTypeScreen(
+            albumId = it.albumId
+        )
+    }
+
+}
