+++ features/photos/photopicker/src/main/kotlin/features/photos/photopicker/presentation/gallerypick/GalleryPhotoPickerState.kt
@@ -1,0 +1,7 @@
+package features.photos.photopicker.presentation.gallerypick
+
+import android.net.Uri
+
+internal data class GalleryPhotoPickerState(
+    val photos: List<Uri> = listOf()
+)
