--- features/photos/src/main/kotlin/features/photos/di/PhotoScreenModule.kt
@@ -1,18 +1,0 @@
-package features.photos.di
-
-import cafe.adriel.voyager.core.registry.screenModule
-import features.photos.presentation.details.PhotoDetailsScreen
-import features.photos.presentation.picktype.PhotoAddTypeScreen
-import platform.navigation.NavigationProvider
-
-val photoScreenModule = screenModule {
-    register<NavigationProvider.Photos.PickPhotoSource> {
-        PhotoAddTypeScreen(
-            albumId = it.albumId
-        )
-    }
-
-    register<NavigationProvider.Photos.PhotoDetails> {
-        PhotoDetailsScreen(it.photoId)
-    }
-}
