--- features/photos/src/main/kotlin/features/photos/presentation/details/PhotoDetailsState.kt
@@ -1,8 +1,0 @@
-package features.photos.presentation.details
-
-import platform.database.models.models.photo.PhotoModel
-
-internal data class PhotoDetailsState(
-    val photo: PhotoModel = PhotoModel.fixture,
-    val isInitialized: Boolean = false
-)
