--- features/photos/src/main/kotlin/features/photos/presentation/gallerypick/GalleryPhotoPickerState.kt
@@ -1,7 +1,0 @@
-package features.photos.presentation.gallerypick
-
-import android.net.Uri
-
-internal data class GalleryPhotoPickerState(
-    val photos: List<Uri> = listOf()
-)
