+++ includes.gradle.kts
@@ -11,4 +11,6 @@ include(":features:albums:home")
 include(":features:albums:create")
 include(":features:albums:details")
 include(":features:albums:shared")
-include(":features:photos")
+include(":features:photos:photopicker")
+include(":features:photos:detail")
+include(":features:photos:shared")
